Squashed commit of the following:
commit 635ddaba3f06b93efbdc8d9d354d5a8cc0231869
Author: tr <jettzeong@gmail.com>
Date:   Wed Jul 12 20:40:44 2023 +0800

    Game now stops when 1 player leaves the room after game start

commit b1b3c2d03a3d1d37ba880fd8b68cc7228570cd35
Author: Hong Jun <zappy.tan1234@gmail.com>
Date:   Wed Jul 12 17:00:31 2023 +0800

    Restructure UI Manager for blitz mode

    Original UI Manager has no Blitz UI elements

commit b9b5daf26c407b320d016086177184dd1f33163b
Author: tr <jettzeong@gmail.com>
Date:   Wed Jul 12 15:30:40 2023 +0800

    Fixed multiple piece instantiation bug

commit 7b650380abade3a552e5772f2d81157715798c62
Author: tr <jettzeong@gmail.com>
Date:   Wed Jul 12 14:06:07 2023 +0800

    Disable joining of room after player starts

    multiplayer is difficult.

commit 84f51983f12c79a44aea536471a58391941b52b3
Author: tr <jettzeong@gmail.com>
Date:   Wed Jul 12 13:58:38 2023 +0800

    Changed LoadScene to LoadSceneAsync

    Somehow fixes the bug that Main Menu scene not loaded when Back button is clicked. Might have something to do with asynchronous processing

commit 58b77748223ab836c4d3a6e2147192cd3ba32b51
Author: tr <jettzeong@gmail.com>
Date:   Wed Jul 12 01:14:07 2023 +0800

    Modified Room Listing Prefab width

commit 02f6085c71e4d42c5d1087c8a531f9e210b294d9
Author: tr <jettzeong@gmail.com>
Date:   Wed Jul 12 01:12:37 2023 +0800

    Added back button for multiplayer panel

commit 5ac1638ed0429fc2df418ee2474cbfe69afe7d7e
Author: tr <jettzeong@gmail.com>
Date:   Wed Jul 12 00:03:58 2023 +0800

    Added ClearCachedList method

commit c91eeadf759e437b08a50a566eca89259092d536
Author: tr <jettzeong@gmail.com>
Date:   Tue Jul 11 23:51:27 2023 +0800

    Updated networking code

commit 19694180c45d11555437adb061b2d72466c58581
Author: tr <jettzeong@gmail.com>
Date:   Tue Jul 11 21:49:38 2023 +0800

    Added scene for multiplayer lobby

commit 5e32bbb007a4fbb1f44cd42ba515b3cbecc99e14
Author: tr <jettzeong@gmail.com>
Date:   Mon Jul 10 12:41:37 2023 +0800

    Custom Game Mode - Bombs now cannot bomb Kings

commit dd8d8467dc4af1332302b3885300a34e0bb6cc3c
Author: tr <jettzeong@gmail.com>
Date:   Mon Jul 10 12:30:27 2023 +0800

    Created InstantiateMine method

commit 22a02e2be14dff0c89312907b9a3b18f6c1e964d
Author: tr <jettzeong@gmail.com>
Date:   Sat Jul 8 11:57:59 2023 +0800

    Deleted unneccesary player selection panel

commit f605c84f1c6d4b5ada0be70001bbeeb164a0f2ac
Author: tr <jettzeong@gmail.com>
Date:   Fri Jul 7 19:17:16 2023 +0800

    Update RoomListingManager.cs

commit d6de6f4836355bee0170cfd99d3b2244399aba15
Author: tr <jettzeong@gmail.com>
Date:   Fri Jul 7 19:17:05 2023 +0800

    Create room button now not interactable if no room name is found

commit 561c8ea3ffccddb40cda6f5c4879b1e634cb189c
Author: tr <jettzeong@gmail.com>
Date:   Fri Jul 7 12:34:51 2023 +0800

    Added InitGame method

    This refreshes the scene every time a new game mode is loaded
    Also, some prefabs in Blitz scene are re-initialized

commit 0bf37edb1cd30cb4aae1a9f37b39a2b37807655c
Author: tr <jettzeong@gmail.com>
Date:   Fri Jul 7 11:55:14 2023 +0800

    Disable both buttons for multiplayer after a player has chosen a playertype

commit d98eade040ed208f14b20c6f3dac6fb12e78bd55
Author: tr <jettzeong@gmail.com>
Date:   Fri Jul 7 11:31:20 2023 +0800

    Check algorithm now factors in promotions

+++ Assets/Scripts/SystemManagement/Game/ModeSelectButton.cs
@@ -4,8 +4,8 @@ using UnityEngine.Assertions;
 
 public class ModeSelectButton : MonoBehaviour
 {
-	[SerializeField] private int minutesPerPlayer;
-	[SerializeField] private int secondsAdded;
+	[SerializeField] private float minutesPerPlayer;
+	[SerializeField] private float secondsAdded;
 	public TMP_Text selectedModeText;
 	public TMP_Text buttonText;
